+++ src/processor/instruction_set.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use log::trace;
 
 use crate::interfaces::Bus as _;
@@ -16,232 +14,225 @@ use InstructionKind::*;
 use MiscInstructionKind::*;
 use StatusRegisterFlag::*;
 
-pub struct InstructionSet {
-    instruction_set: HashMap<Opcode, Instruction>,
-}
+pub struct InstructionSet;
 
 impl InstructionSet {
-    #[rustfmt::skip]
     pub fn new_legal_opcode_set() -> Self {
-        let mut instruction_set = HashMap::new();
-
-        let instructions = [
-            // Transfer instructions
-            Instruction {name: "LDA", opcode: 0xA9, instruction: InternalExecOnMemoryData(lda), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "LDA", opcode: 0xA5, instruction: InternalExecOnMemoryData(lda), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "LDA", opcode: 0xB5, instruction: InternalExecOnMemoryData(lda), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDA", opcode: 0xAD, instruction: InternalExecOnMemoryData(lda), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDA", opcode: 0xBD, instruction: InternalExecOnMemoryData(lda), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "LDA", opcode: 0xB9, instruction: InternalExecOnMemoryData(lda), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "LDA", opcode: 0xA1, instruction: InternalExecOnMemoryData(lda), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "LDA", opcode: 0xB1, instruction: InternalExecOnMemoryData(lda), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            Instruction {name: "LDX", opcode: 0xA2, instruction: InternalExecOnMemoryData(ldx), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "LDX", opcode: 0xA6, instruction: InternalExecOnMemoryData(ldx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "LDX", opcode: 0xB6, instruction: InternalExecOnMemoryData(ldx), addressing_mode: ZeroPageY, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDX", opcode: 0xAE, instruction: InternalExecOnMemoryData(ldx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDX", opcode: 0xBE, instruction: InternalExecOnMemoryData(ldx), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-
-            Instruction {name: "LDY", opcode: 0xA0, instruction: InternalExecOnMemoryData(ldy), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "LDY", opcode: 0xA4, instruction: InternalExecOnMemoryData(ldy), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "LDY", opcode: 0xB4, instruction: InternalExecOnMemoryData(ldy), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDY", opcode: 0xAC, instruction: InternalExecOnMemoryData(ldy), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "LDY", opcode: 0xBC, instruction: InternalExecOnMemoryData(ldy), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-
-            Instruction {name: "STA", opcode: 0x85, instruction: StoreOp(sta), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x95, instruction: StoreOp(sta), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x8D, instruction: StoreOp(sta), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x9D, instruction: StoreOp(sta), addressing_mode: AbsoluteX, bytes: 3, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x99, instruction: StoreOp(sta), addressing_mode: AbsoluteY, bytes: 3, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x81, instruction: StoreOp(sta), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "STA", opcode: 0x91, instruction: StoreOp(sta), addressing_mode: IndirectY, bytes: 2, cycles: 6, page_crossing_cost: 0},
-
-            Instruction {name: "STX", opcode: 0x86, instruction: StoreOp(stx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "STX", opcode: 0x96, instruction: StoreOp(stx), addressing_mode: ZeroPageY, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "STX", opcode: 0x8E, instruction: StoreOp(stx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-
-            Instruction {name: "STY", opcode: 0x84, instruction: StoreOp(sty), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "STY", opcode: 0x94, instruction: StoreOp(sty), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "STY", opcode: 0x8C, instruction: StoreOp(sty), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-
-            Instruction {name: "TAX", opcode: 0xAA, instruction: SingleByte(tax), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "TAY", opcode: 0xA8, instruction: SingleByte(tay), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "TSX", opcode: 0xBA, instruction: SingleByte(tsx), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "TXA", opcode: 0x8A, instruction: SingleByte(txa), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "TXS", opcode: 0x9A, instruction: SingleByte(txs), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "TYA", opcode: 0x98, instruction: SingleByte(tya), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            // Stack instructions
-            Instruction {name: "PHA", opcode: 0x48, instruction: Misc(Push(pha)), addressing_mode: Implied, bytes: 1, cycles: 3, page_crossing_cost: 0},
-
-            Instruction {name: "PHP", opcode: 0x08, instruction: Misc(Push(php)), addressing_mode: Implied, bytes: 1, cycles: 3, page_crossing_cost: 0},
-
-            Instruction {name: "PLA", opcode: 0x68, instruction: Misc(Pull(pla)), addressing_mode: Implied, bytes: 1, cycles: 4, page_crossing_cost: 0},
-
-            Instruction {name: "PLP", opcode: 0x28, instruction: Misc(Pull(plp)), addressing_mode: Implied, bytes: 1, cycles: 4, page_crossing_cost: 0},
-
-            // Decrements and increments
-            Instruction {name: "DEC", opcode: 0xC6, instruction: ReadModifyWrite(dec), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "DEC", opcode: 0xD6, instruction: ReadModifyWrite(dec), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "DEC", opcode: 0xCE, instruction: ReadModifyWrite(dec), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "DEC", opcode: 0xDE, instruction: ReadModifyWrite(dec), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "DEX", opcode: 0xCA, instruction: SingleByte(dex), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "DEY", opcode: 0x88, instruction: SingleByte(dey), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "INC", opcode: 0xE6, instruction: ReadModifyWrite(inc), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "INC", opcode: 0xF6, instruction: ReadModifyWrite(inc), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "INC", opcode: 0xEE, instruction: ReadModifyWrite(inc), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "INC", opcode: 0xFE, instruction: ReadModifyWrite(inc), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "INX", opcode: 0xE8, instruction: SingleByte(inx), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            Instruction {name: "INY", opcode: 0xC8, instruction: SingleByte(iny), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            // Arithmetic operations
-            Instruction {name: "ADC", opcode: 0x69, instruction: InternalExecOnMemoryData(adc), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "ADC", opcode: 0x65, instruction: InternalExecOnMemoryData(adc), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "ADC", opcode: 0x75, instruction: InternalExecOnMemoryData(adc), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "ADC", opcode: 0x6D, instruction: InternalExecOnMemoryData(adc), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "ADC", opcode: 0x7D, instruction: InternalExecOnMemoryData(adc), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "ADC", opcode: 0x79, instruction: InternalExecOnMemoryData(adc), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "ADC", opcode: 0x61, instruction: InternalExecOnMemoryData(adc), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ADC", opcode: 0x71, instruction: InternalExecOnMemoryData(adc), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            Instruction {name: "SBC", opcode: 0xE9, instruction: InternalExecOnMemoryData(sbc), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "SBC", opcode: 0xE5, instruction: InternalExecOnMemoryData(sbc), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "SBC", opcode: 0xF5, instruction: InternalExecOnMemoryData(sbc), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "SBC", opcode: 0xED, instruction: InternalExecOnMemoryData(sbc), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "SBC", opcode: 0xFD, instruction: InternalExecOnMemoryData(sbc), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "SBC", opcode: 0xF9, instruction: InternalExecOnMemoryData(sbc), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "SBC", opcode: 0xE1, instruction: InternalExecOnMemoryData(sbc), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "SBC", opcode: 0xF1, instruction: InternalExecOnMemoryData(sbc), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            // Logical operations
-            Instruction {name: "AND", opcode: 0x29, instruction: InternalExecOnMemoryData(and), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "AND", opcode: 0x25, instruction: InternalExecOnMemoryData(and), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "AND", opcode: 0x35, instruction: InternalExecOnMemoryData(and), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "AND", opcode: 0x2D, instruction: InternalExecOnMemoryData(and), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "AND", opcode: 0x3D, instruction: InternalExecOnMemoryData(and), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "AND", opcode: 0x39, instruction: InternalExecOnMemoryData(and), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "AND", opcode: 0x21, instruction: InternalExecOnMemoryData(and), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "AND", opcode: 0x31, instruction: InternalExecOnMemoryData(and), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            Instruction {name: "EOR", opcode: 0x49, instruction: InternalExecOnMemoryData(eor), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "EOR", opcode: 0x45, instruction: InternalExecOnMemoryData(eor), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "EOR", opcode: 0x55, instruction: InternalExecOnMemoryData(eor), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "EOR", opcode: 0x4D, instruction: InternalExecOnMemoryData(eor), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "EOR", opcode: 0x5D, instruction: InternalExecOnMemoryData(eor), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "EOR", opcode: 0x59, instruction: InternalExecOnMemoryData(eor), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "EOR", opcode: 0x41, instruction: InternalExecOnMemoryData(eor), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "EOR", opcode: 0x51, instruction: InternalExecOnMemoryData(eor), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            Instruction {name: "ORA", opcode: 0x09, instruction: InternalExecOnMemoryData(ora), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "ORA", opcode: 0x05, instruction: InternalExecOnMemoryData(ora), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "ORA", opcode: 0x15, instruction: InternalExecOnMemoryData(ora), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "ORA", opcode: 0x0D, instruction: InternalExecOnMemoryData(ora), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "ORA", opcode: 0x1D, instruction: InternalExecOnMemoryData(ora), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "ORA", opcode: 0x19, instruction: InternalExecOnMemoryData(ora), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "ORA", opcode: 0x01, instruction: InternalExecOnMemoryData(ora), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ORA", opcode: 0x11, instruction: InternalExecOnMemoryData(ora), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            // Shift and rotation instructions
-            Instruction {name: "ASL", opcode: 0x0A, instruction: SingleByte(asl_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "ASL", opcode: 0x06, instruction: ReadModifyWrite(asl), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "ASL", opcode: 0x16, instruction: ReadModifyWrite(asl), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ASL", opcode: 0x0E, instruction: ReadModifyWrite(asl), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ASL", opcode: 0x1E, instruction: ReadModifyWrite(asl), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "LSR", opcode: 0x4A, instruction: SingleByte(lsr_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "LSR", opcode: 0x46, instruction: ReadModifyWrite(lsr), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "LSR", opcode: 0x56, instruction: ReadModifyWrite(lsr), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "LSR", opcode: 0x4E, instruction: ReadModifyWrite(lsr), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "LSR", opcode: 0x5E, instruction: ReadModifyWrite(lsr), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "ROL", opcode: 0x2A, instruction: SingleByte(rol_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "ROL", opcode: 0x26, instruction: ReadModifyWrite(rol), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "ROL", opcode: 0x36, instruction: ReadModifyWrite(rol), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ROL", opcode: 0x2E, instruction: ReadModifyWrite(rol), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ROL", opcode: 0x3E, instruction: ReadModifyWrite(rol), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "ROR", opcode: 0x6A, instruction: SingleByte(ror_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "ROR", opcode: 0x66, instruction: ReadModifyWrite(ror), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
-            Instruction {name: "ROR", opcode: 0x76, instruction: ReadModifyWrite(ror), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ROR", opcode: 0x6E, instruction: ReadModifyWrite(ror), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "ROR", opcode: 0x7E, instruction: ReadModifyWrite(ror), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
-
-            // Flag instructions
-            Instruction {name: "CLC", opcode: 0x18, instruction: SingleByte(clc), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CLD", opcode: 0xD8, instruction: SingleByte(cld), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CLI", opcode: 0x58, instruction: SingleByte(cli), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CLV", opcode: 0xB8, instruction: SingleByte(clv), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "SEC", opcode: 0x38, instruction: SingleByte(sec), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "SED", opcode: 0xF8, instruction: SingleByte(sed), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "SEI", opcode: 0x78, instruction: SingleByte(sei), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-
-            // Comparaisons
-            Instruction {name: "CMP", opcode: 0xC9, instruction: InternalExecOnMemoryData(cmp), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CMP", opcode: 0xC5, instruction: InternalExecOnMemoryData(cmp), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "CMP", opcode: 0xD5, instruction: InternalExecOnMemoryData(cmp), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "CMP", opcode: 0xCD, instruction: InternalExecOnMemoryData(cmp), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-            Instruction {name: "CMP", opcode: 0xDD, instruction: InternalExecOnMemoryData(cmp), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "CMP", opcode: 0xD9, instruction: InternalExecOnMemoryData(cmp), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
-            Instruction {name: "CMP", opcode: 0xC1, instruction: InternalExecOnMemoryData(cmp), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
-            Instruction {name: "CMP", opcode: 0xD1, instruction: InternalExecOnMemoryData(cmp), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
-
-            Instruction {name: "CPX", opcode: 0xE0, instruction: InternalExecOnMemoryData(cpx), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CPX", opcode: 0xE4, instruction: InternalExecOnMemoryData(cpx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "CPX", opcode: 0xEC, instruction: InternalExecOnMemoryData(cpx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-
-            Instruction {name: "CPY", opcode: 0xC0, instruction: InternalExecOnMemoryData(cpy), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "CPY", opcode: 0xC4, instruction: InternalExecOnMemoryData(cpy), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "CPY", opcode: 0xCC, instruction: InternalExecOnMemoryData(cpy), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-
-            // Conditional branch instructions
-            Instruction {name: "BCC", opcode: 0x90, instruction: Misc(Branch(bcc)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BCS", opcode: 0xB0, instruction: Misc(Branch(bcs)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BEQ", opcode: 0xF0, instruction: Misc(Branch(beq)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BMI", opcode: 0x30, instruction: Misc(Branch(bmi)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BNE", opcode: 0xD0, instruction: Misc(Branch(bne)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BPL", opcode: 0x10, instruction: Misc(Branch(bpl)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BVC", opcode: 0x50, instruction: Misc(Branch(bvc)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-            Instruction {name: "BVS", opcode: 0x70, instruction: Misc(Branch(bvs)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
-
-            // Jumps and subroutines
-            Instruction {name: "JMP", opcode: 0x4C, instruction: Misc(Jump(jmp)), addressing_mode: Absolute, bytes: 3, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "JMP", opcode: 0x6C, instruction: Misc(Jump(jmp)), addressing_mode: Indirect, bytes: 3, cycles: 5, page_crossing_cost: 0},
-
-            Instruction {name: "JSR", opcode: 0x20, instruction: Misc(Call(jsr)), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
-
-            Instruction {name: "RTI", opcode: 0x40, instruction: Misc(ReturnFromInterrupt(rti)), addressing_mode: Implied, bytes: 1, cycles: 6, page_crossing_cost: 0},
-
-            // Interrupts
-            Instruction {name: "BRK", opcode: 0x00, instruction: Misc(HardwareInterrupt(brk)), addressing_mode: Implied, bytes: 1, cycles: 7, page_crossing_cost: 0},
-
-            Instruction {name: "RTS", opcode: 0x60, instruction: Misc(Return(rts)), addressing_mode: Implied, bytes: 1, cycles: 6, page_crossing_cost: 0},
-
-            // Other
-            Instruction {name: "BIT", opcode: 0x24, instruction: InternalExecOnMemoryData(bit), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
-            Instruction {name: "BIT", opcode: 0x2C, instruction: InternalExecOnMemoryData(bit), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
-
-            Instruction {name: "NOP", opcode: 0xEA, instruction: SingleByte(nop), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
-        ];
-
-        for instruction in instructions {
-            instruction_set.insert(instruction.opcode, instruction);
-        }
-
-        Self { instruction_set }
+        Self {}
     }
 
+    #[rustfmt::skip]
     pub fn lookup(&self, opcode: Opcode) -> Option<Instruction> {
-        self.instruction_set.get(&opcode).cloned()
+        let instruction = match opcode {
+            // Transfer instructions
+            0xA9 => Instruction {name: "LDA", opcode: 0xA9, instruction: InternalExecOnMemoryData(lda), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xA5 => Instruction {name: "LDA", opcode: 0xA5, instruction: InternalExecOnMemoryData(lda), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xB5 => Instruction {name: "LDA", opcode: 0xB5, instruction: InternalExecOnMemoryData(lda), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0xAD => Instruction {name: "LDA", opcode: 0xAD, instruction: InternalExecOnMemoryData(lda), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0xBD => Instruction {name: "LDA", opcode: 0xBD, instruction: InternalExecOnMemoryData(lda), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xB9 => Instruction {name: "LDA", opcode: 0xB9, instruction: InternalExecOnMemoryData(lda), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xA1 => Instruction {name: "LDA", opcode: 0xA1, instruction: InternalExecOnMemoryData(lda), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0xB1 => Instruction {name: "LDA", opcode: 0xB1, instruction: InternalExecOnMemoryData(lda), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            0xA2 => Instruction {name: "LDX", opcode: 0xA2, instruction: InternalExecOnMemoryData(ldx), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xA6 => Instruction {name: "LDX", opcode: 0xA6, instruction: InternalExecOnMemoryData(ldx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xB6 => Instruction {name: "LDX", opcode: 0xB6, instruction: InternalExecOnMemoryData(ldx), addressing_mode: ZeroPageY, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0xAE => Instruction {name: "LDX", opcode: 0xAE, instruction: InternalExecOnMemoryData(ldx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0xBE => Instruction {name: "LDX", opcode: 0xBE, instruction: InternalExecOnMemoryData(ldx), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+
+            0xA0 => Instruction {name: "LDY", opcode: 0xA0, instruction: InternalExecOnMemoryData(ldy), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xA4 => Instruction {name: "LDY", opcode: 0xA4, instruction: InternalExecOnMemoryData(ldy), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xB4 => Instruction {name: "LDY", opcode: 0xB4, instruction: InternalExecOnMemoryData(ldy), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0xAC => Instruction {name: "LDY", opcode: 0xAC, instruction: InternalExecOnMemoryData(ldy), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0xBC => Instruction {name: "LDY", opcode: 0xBC, instruction: InternalExecOnMemoryData(ldy), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+
+            0x85 => Instruction {name: "STA", opcode: 0x85, instruction: StoreOp(sta), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x95 => Instruction {name: "STA", opcode: 0x95, instruction: StoreOp(sta), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x8D => Instruction {name: "STA", opcode: 0x8D, instruction: StoreOp(sta), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0x9D => Instruction {name: "STA", opcode: 0x9D, instruction: StoreOp(sta), addressing_mode: AbsoluteX, bytes: 3, cycles: 5, page_crossing_cost: 0},
+            0x99 => Instruction {name: "STA", opcode: 0x99, instruction: StoreOp(sta), addressing_mode: AbsoluteY, bytes: 3, cycles: 5, page_crossing_cost: 0},
+            0x81 => Instruction {name: "STA", opcode: 0x81, instruction: StoreOp(sta), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x91 => Instruction {name: "STA", opcode: 0x91, instruction: StoreOp(sta), addressing_mode: IndirectY, bytes: 2, cycles: 6, page_crossing_cost: 0},
+
+            0x86 => Instruction {name: "STX", opcode: 0x86, instruction: StoreOp(stx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x96 => Instruction {name: "STX", opcode: 0x96, instruction: StoreOp(stx), addressing_mode: ZeroPageY, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x8E => Instruction {name: "STX", opcode: 0x8E, instruction: StoreOp(stx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+
+            0x84 => Instruction {name: "STY", opcode: 0x84, instruction: StoreOp(sty), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x94 => Instruction {name: "STY", opcode: 0x94, instruction: StoreOp(sty), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x8C => Instruction {name: "STY", opcode: 0x8C, instruction: StoreOp(sty), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+
+            0xAA => Instruction {name: "TAX", opcode: 0xAA, instruction: SingleByte(tax), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0xA8 => Instruction {name: "TAY", opcode: 0xA8, instruction: SingleByte(tay), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0xBA => Instruction {name: "TSX", opcode: 0xBA, instruction: SingleByte(tsx), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0x8A => Instruction {name: "TXA", opcode: 0x8A, instruction: SingleByte(txa), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0x9A => Instruction {name: "TXS", opcode: 0x9A, instruction: SingleByte(txs), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0x98 => Instruction {name: "TYA", opcode: 0x98, instruction: SingleByte(tya), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            // Stack instructions
+            0x48 => Instruction {name: "PHA", opcode: 0x48, instruction: Misc(Push(pha)), addressing_mode: Implied, bytes: 1, cycles: 3, page_crossing_cost: 0},
+
+            0x08 => Instruction {name: "PHP", opcode: 0x08, instruction: Misc(Push(php)), addressing_mode: Implied, bytes: 1, cycles: 3, page_crossing_cost: 0},
+
+            0x68 => Instruction {name: "PLA", opcode: 0x68, instruction: Misc(Pull(pla)), addressing_mode: Implied, bytes: 1, cycles: 4, page_crossing_cost: 0},
+
+            0x28 => Instruction {name: "PLP", opcode: 0x28, instruction: Misc(Pull(plp)), addressing_mode: Implied, bytes: 1, cycles: 4, page_crossing_cost: 0},
+
+            // Decrements and increments
+            0xC6 => Instruction {name: "DEC", opcode: 0xC6, instruction: ReadModifyWrite(dec), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0xD6 => Instruction {name: "DEC", opcode: 0xD6, instruction: ReadModifyWrite(dec), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0xCE => Instruction {name: "DEC", opcode: 0xCE, instruction: ReadModifyWrite(dec), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0xDE => Instruction {name: "DEC", opcode: 0xDE, instruction: ReadModifyWrite(dec), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            0xCA => Instruction {name: "DEX", opcode: 0xCA, instruction: SingleByte(dex), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0x88 => Instruction {name: "DEY", opcode: 0x88, instruction: SingleByte(dey), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0xE6 => Instruction {name: "INC", opcode: 0xE6, instruction: ReadModifyWrite(inc), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0xF6 => Instruction {name: "INC", opcode: 0xF6, instruction: ReadModifyWrite(inc), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0xEE => Instruction {name: "INC", opcode: 0xEE, instruction: ReadModifyWrite(inc), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0xFE => Instruction {name: "INC", opcode: 0xFE, instruction: ReadModifyWrite(inc), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            0xE8 => Instruction {name: "INX", opcode: 0xE8, instruction: SingleByte(inx), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            0xC8 => Instruction {name: "INY", opcode: 0xC8, instruction: SingleByte(iny), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            // Arithmetic operations
+            0x69 => Instruction {name: "ADC", opcode: 0x69, instruction: InternalExecOnMemoryData(adc), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x65 => Instruction {name: "ADC", opcode: 0x65, instruction: InternalExecOnMemoryData(adc), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x75 => Instruction {name: "ADC", opcode: 0x75, instruction: InternalExecOnMemoryData(adc), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x6D => Instruction {name: "ADC", opcode: 0x6D, instruction: InternalExecOnMemoryData(adc), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0x7D => Instruction {name: "ADC", opcode: 0x7D, instruction: InternalExecOnMemoryData(adc), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x79 => Instruction {name: "ADC", opcode: 0x79, instruction: InternalExecOnMemoryData(adc), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x61 => Instruction {name: "ADC", opcode: 0x61, instruction: InternalExecOnMemoryData(adc), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x71 => Instruction {name: "ADC", opcode: 0x71, instruction: InternalExecOnMemoryData(adc), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            0xE9 => Instruction {name: "SBC", opcode: 0xE9, instruction: InternalExecOnMemoryData(sbc), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xE5 => Instruction {name: "SBC", opcode: 0xE5, instruction: InternalExecOnMemoryData(sbc), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xF5 => Instruction {name: "SBC", opcode: 0xF5, instruction: InternalExecOnMemoryData(sbc), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0xED => Instruction {name: "SBC", opcode: 0xED, instruction: InternalExecOnMemoryData(sbc), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0xFD => Instruction {name: "SBC", opcode: 0xFD, instruction: InternalExecOnMemoryData(sbc), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xF9 => Instruction {name: "SBC", opcode: 0xF9, instruction: InternalExecOnMemoryData(sbc), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xE1 => Instruction {name: "SBC", opcode: 0xE1, instruction: InternalExecOnMemoryData(sbc), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0xF1 => Instruction {name: "SBC", opcode: 0xF1, instruction: InternalExecOnMemoryData(sbc), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            // Logical operations
+            0x29 => Instruction {name: "AND", opcode: 0x29, instruction: InternalExecOnMemoryData(and), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x25 => Instruction {name: "AND", opcode: 0x25, instruction: InternalExecOnMemoryData(and), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x35 => Instruction {name: "AND", opcode: 0x35, instruction: InternalExecOnMemoryData(and), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x2D => Instruction {name: "AND", opcode: 0x2D, instruction: InternalExecOnMemoryData(and), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0x3D => Instruction {name: "AND", opcode: 0x3D, instruction: InternalExecOnMemoryData(and), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x39 => Instruction {name: "AND", opcode: 0x39, instruction: InternalExecOnMemoryData(and), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x21 => Instruction {name: "AND", opcode: 0x21, instruction: InternalExecOnMemoryData(and), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x31 => Instruction {name: "AND", opcode: 0x31, instruction: InternalExecOnMemoryData(and), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            0x49 => Instruction {name: "EOR", opcode: 0x49, instruction: InternalExecOnMemoryData(eor), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x45 => Instruction {name: "EOR", opcode: 0x45, instruction: InternalExecOnMemoryData(eor), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x55 => Instruction {name: "EOR", opcode: 0x55, instruction: InternalExecOnMemoryData(eor), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x4D => Instruction {name: "EOR", opcode: 0x4D, instruction: InternalExecOnMemoryData(eor), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0x5D => Instruction {name: "EOR", opcode: 0x5D, instruction: InternalExecOnMemoryData(eor), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x59 => Instruction {name: "EOR", opcode: 0x59, instruction: InternalExecOnMemoryData(eor), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x41 => Instruction {name: "EOR", opcode: 0x41, instruction: InternalExecOnMemoryData(eor), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x51 => Instruction {name: "EOR", opcode: 0x51, instruction: InternalExecOnMemoryData(eor), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            0x09 => Instruction {name: "ORA", opcode: 0x09, instruction: InternalExecOnMemoryData(ora), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x05 => Instruction {name: "ORA", opcode: 0x05, instruction: InternalExecOnMemoryData(ora), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x15 => Instruction {name: "ORA", opcode: 0x15, instruction: InternalExecOnMemoryData(ora), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0x0D => Instruction {name: "ORA", opcode: 0x0D, instruction: InternalExecOnMemoryData(ora), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0x1D => Instruction {name: "ORA", opcode: 0x1D, instruction: InternalExecOnMemoryData(ora), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x19 => Instruction {name: "ORA", opcode: 0x19, instruction: InternalExecOnMemoryData(ora), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0x01 => Instruction {name: "ORA", opcode: 0x01, instruction: InternalExecOnMemoryData(ora), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x11 => Instruction {name: "ORA", opcode: 0x11, instruction: InternalExecOnMemoryData(ora), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            // Shift and rotation instructions
+            0x0A => Instruction {name: "ASL", opcode: 0x0A, instruction: SingleByte(asl_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x06 => Instruction {name: "ASL", opcode: 0x06, instruction: ReadModifyWrite(asl), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0x16 => Instruction {name: "ASL", opcode: 0x16, instruction: ReadModifyWrite(asl), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x0E => Instruction {name: "ASL", opcode: 0x0E, instruction: ReadModifyWrite(asl), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0x1E => Instruction {name: "ASL", opcode: 0x1E, instruction: ReadModifyWrite(asl), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            0x4A => Instruction {name: "LSR", opcode: 0x4A, instruction: SingleByte(lsr_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x46 => Instruction {name: "LSR", opcode: 0x46, instruction: ReadModifyWrite(lsr), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0x56 => Instruction {name: "LSR", opcode: 0x56, instruction: ReadModifyWrite(lsr), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x4E => Instruction {name: "LSR", opcode: 0x4E, instruction: ReadModifyWrite(lsr), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0x5E => Instruction {name: "LSR", opcode: 0x5E, instruction: ReadModifyWrite(lsr), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            0x2A => Instruction {name: "ROL", opcode: 0x2A, instruction: SingleByte(rol_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x26 => Instruction {name: "ROL", opcode: 0x26, instruction: ReadModifyWrite(rol), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0x36 => Instruction {name: "ROL", opcode: 0x36, instruction: ReadModifyWrite(rol), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x2E => Instruction {name: "ROL", opcode: 0x2E, instruction: ReadModifyWrite(rol), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0x3E => Instruction {name: "ROL", opcode: 0x3E, instruction: ReadModifyWrite(rol), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            0x6A => Instruction {name: "ROR", opcode: 0x6A, instruction: SingleByte(ror_acc), addressing_mode: Accumulator, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x66 => Instruction {name: "ROR", opcode: 0x66, instruction: ReadModifyWrite(ror), addressing_mode: ZeroPage, bytes: 2, cycles: 5, page_crossing_cost: 0},
+            0x76 => Instruction {name: "ROR", opcode: 0x76, instruction: ReadModifyWrite(ror), addressing_mode: ZeroPageX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0x6E => Instruction {name: "ROR", opcode: 0x6E, instruction: ReadModifyWrite(ror), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+            0x7E => Instruction {name: "ROR", opcode: 0x7E, instruction: ReadModifyWrite(ror), addressing_mode: AbsoluteX, bytes: 3, cycles: 7, page_crossing_cost: 0},
+
+            // Flag instructions
+            0x18 => Instruction {name: "CLC", opcode: 0x18, instruction: SingleByte(clc), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0xD8 => Instruction {name: "CLD", opcode: 0xD8, instruction: SingleByte(cld), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x58 => Instruction {name: "CLI", opcode: 0x58, instruction: SingleByte(cli), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0xB8 => Instruction {name: "CLV", opcode: 0xB8, instruction: SingleByte(clv), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x38 => Instruction {name: "SEC", opcode: 0x38, instruction: SingleByte(sec), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0xF8 => Instruction {name: "SED", opcode: 0xF8, instruction: SingleByte(sed), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+            0x78 => Instruction {name: "SEI", opcode: 0x78, instruction: SingleByte(sei), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            // Comparaisons
+            0xC9 => Instruction {name: "CMP", opcode: 0xC9, instruction: InternalExecOnMemoryData(cmp), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xC5 => Instruction {name: "CMP", opcode: 0xC5, instruction: InternalExecOnMemoryData(cmp), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xD5 => Instruction {name: "CMP", opcode: 0xD5, instruction: InternalExecOnMemoryData(cmp), addressing_mode: ZeroPageX, bytes: 2, cycles: 4, page_crossing_cost: 0},
+            0xCD => Instruction {name: "CMP", opcode: 0xCD, instruction: InternalExecOnMemoryData(cmp), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+            0xDD => Instruction {name: "CMP", opcode: 0xDD, instruction: InternalExecOnMemoryData(cmp), addressing_mode: AbsoluteX, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xD9 => Instruction {name: "CMP", opcode: 0xD9, instruction: InternalExecOnMemoryData(cmp), addressing_mode: AbsoluteY, bytes: 3, cycles: 4, page_crossing_cost: 1},
+            0xC1 => Instruction {name: "CMP", opcode: 0xC1, instruction: InternalExecOnMemoryData(cmp), addressing_mode: IndirectX, bytes: 2, cycles: 6, page_crossing_cost: 0},
+            0xD1 => Instruction {name: "CMP", opcode: 0xD1, instruction: InternalExecOnMemoryData(cmp), addressing_mode: IndirectY, bytes: 2, cycles: 5, page_crossing_cost: 1},
+
+            0xE0 => Instruction {name: "CPX", opcode: 0xE0, instruction: InternalExecOnMemoryData(cpx), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xE4 => Instruction {name: "CPX", opcode: 0xE4, instruction: InternalExecOnMemoryData(cpx), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xEC => Instruction {name: "CPX", opcode: 0xEC, instruction: InternalExecOnMemoryData(cpx), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+
+            0xC0 => Instruction {name: "CPY", opcode: 0xC0, instruction: InternalExecOnMemoryData(cpy), addressing_mode: Immediate, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xC4 => Instruction {name: "CPY", opcode: 0xC4, instruction: InternalExecOnMemoryData(cpy), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0xCC => Instruction {name: "CPY", opcode: 0xCC, instruction: InternalExecOnMemoryData(cpy), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+
+            // Conditional branch instructions
+            0x90 => Instruction {name: "BCC", opcode: 0x90, instruction: Misc(Branch(bcc)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xB0 => Instruction {name: "BCS", opcode: 0xB0, instruction: Misc(Branch(bcs)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xF0 => Instruction {name: "BEQ", opcode: 0xF0, instruction: Misc(Branch(beq)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x30 => Instruction {name: "BMI", opcode: 0x30, instruction: Misc(Branch(bmi)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0xD0 => Instruction {name: "BNE", opcode: 0xD0, instruction: Misc(Branch(bne)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x10 => Instruction {name: "BPL", opcode: 0x10, instruction: Misc(Branch(bpl)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x50 => Instruction {name: "BVC", opcode: 0x50, instruction: Misc(Branch(bvc)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+            0x70 => Instruction {name: "BVS", opcode: 0x70, instruction: Misc(Branch(bvs)), addressing_mode: Relative, bytes: 2, cycles: 2, page_crossing_cost: 0},
+
+            // Jumps and subroutines
+            0x4C => Instruction {name: "JMP", opcode: 0x4C, instruction: Misc(Jump(jmp)), addressing_mode: Absolute, bytes: 3, cycles: 3, page_crossing_cost: 0},
+            0x6C => Instruction {name: "JMP", opcode: 0x6C, instruction: Misc(Jump(jmp)), addressing_mode: Indirect, bytes: 3, cycles: 5, page_crossing_cost: 0},
+
+            0x20 => Instruction {name: "JSR", opcode: 0x20, instruction: Misc(Call(jsr)), addressing_mode: Absolute, bytes: 3, cycles: 6, page_crossing_cost: 0},
+
+            0x40 => Instruction {name: "RTI", opcode: 0x40, instruction: Misc(ReturnFromInterrupt(rti)), addressing_mode: Implied, bytes: 1, cycles: 6, page_crossing_cost: 0},
+
+            // Interrupts
+            0x00 => Instruction {name: "BRK", opcode: 0x00, instruction: Misc(HardwareInterrupt(brk)), addressing_mode: Implied, bytes: 1, cycles: 7, page_crossing_cost: 0},
+
+            0x60 => Instruction {name: "RTS", opcode: 0x60, instruction: Misc(Return(rts)), addressing_mode: Implied, bytes: 1, cycles: 6, page_crossing_cost: 0},
+
+            // Other
+            0x24 => Instruction {name: "BIT", opcode: 0x24, instruction: InternalExecOnMemoryData(bit), addressing_mode: ZeroPage, bytes: 2, cycles: 3, page_crossing_cost: 0},
+            0x2C => Instruction {name: "BIT", opcode: 0x2C, instruction: InternalExecOnMemoryData(bit), addressing_mode: Absolute, bytes: 3, cycles: 4, page_crossing_cost: 0},
+
+            0xEA => Instruction {name: "NOP", opcode: 0xEA, instruction: SingleByte(nop), addressing_mode: Implied, bytes: 1, cycles: 2, page_crossing_cost: 0},
+
+            _ => return None,
+        };
+        Some(instruction)
     }
 }